Move functional test setup into BaseStripeFunctionalTest
